+++ argopy/tests/test_utilities.py
@@ -9,17 +9,7 @@ import types
 
 import argopy
 from argopy.utilities import (
-    load_dict,
-    mapp_dict,
-    list_multiprofile_file_variables,
-    check_gdac_path,
-    isconnected,
-    urlhaskeyword,
-    isalive,
-    isAPIconnected,
-    erddap_ds_exists,
     linear_interpolation_remap,
-    Chunker,
     format_oneline,
     wmo2box,
     modified_environ,
@@ -52,33 +42,6 @@ def test_show_versions(conda):
     f = io.StringIO()
     argopy.show_versions(file=f, conda=conda)
     assert "SYSTEM" in f.getvalue()
-
-
-@requires_gdac
-def test_clear_cache():
-    ftproot, flist = argopy.tutorial.open_dataset("gdac")
-    with tempfile.TemporaryDirectory() as cachedir:
-        with argopy.set_options(cachedir=cachedir):
-            loader = ArgoDataFetcher(src="gdac", ftp=ftproot, cache=True).profile(2902696, 12)
-            loader.to_xarray()
-            argopy.clear_cache()
-            assert os.path.exists(cachedir) is True
-            assert len(os.listdir(cachedir)) == 0
-
-
-@requires_gdac
-def test_lscache():
-    ftproot, flist = argopy.tutorial.open_dataset("gdac")
-    with tempfile.TemporaryDirectory() as cachedir:
-        with argopy.set_options(cachedir=cachedir):
-            loader = ArgoDataFetcher(src="gdac", ftp=ftproot, cache=True).profile(2902696, 12)
-            loader.to_xarray()
-            result = argopy.utilities.lscache(cache_path=cachedir, prt=True)
-            assert isinstance(result, str)
-
-            result = argopy.utilities.lscache(cache_path=cachedir, prt=False)
-            assert isinstance(result, pd.DataFrame)
-
 
 
 class Test_linear_interpolation_remap:
@@ -148,193 +111,6 @@ class Test_linear_interpolation_remap:
                 z_dim="N_LEVELS",
                 z_regridded_dim="Z_LEVELS",
             )
-
-
-class Test_Chunker:
-    @pytest.fixture(autouse=True)
-    def create_data(self):
-        self.WMO = [
-            6902766,
-            6902772,
-            6902914,
-            6902746,
-            6902916,
-            6902915,
-            6902757,
-            6902771,
-        ]
-        self.BOX3d = [0, 20, 40, 60, 0, 1000]
-        self.BOX4d = [0, 20, 40, 60, 0, 1000, "2001-01", "2001-6"]
-
-    def test_InvalidFetcherAccessPoint(self):
-        with pytest.raises(InvalidFetcherAccessPoint):
-            Chunker({"invalid": self.WMO})
-
-    def test_invalid_chunks(self):
-        with pytest.raises(ValueError):
-            Chunker({"box": self.BOX3d}, chunks='toto')
-
-    def test_invalid_chunksize(self):
-        with pytest.raises(ValueError):
-            Chunker({"box": self.BOX3d}, chunksize='toto')
-
-    def test_chunk_wmo(self):
-        C = Chunker({"wmo": self.WMO})
-        assert all(
-            [all(isinstance(x, int) for x in chunk) for chunk in C.fit_transform()]
-        )
-
-        C = Chunker({"wmo": self.WMO}, chunks="auto")
-        assert all(
-            [all(isinstance(x, int) for x in chunk) for chunk in C.fit_transform()]
-        )
-
-        C = Chunker({"wmo": self.WMO}, chunks={"wmo": 1})
-        assert all(
-            [all(isinstance(x, int) for x in chunk) for chunk in C.fit_transform()]
-        )
-        assert len(C.fit_transform()) == 1
-
-        with pytest.raises(ValueError):
-            Chunker({"wmo": self.WMO}, chunks=["wmo", 1])
-
-        C = Chunker({"wmo": self.WMO})
-        assert isinstance(C.this_chunker, types.FunctionType) or isinstance(
-            C.this_chunker, types.MethodType
-        )
-
-    def test_chunk_box3d(self):
-        C = Chunker({"box": self.BOX3d})
-        assert all([is_box(chunk) for chunk in C.fit_transform()])
-
-        C = Chunker({"box": self.BOX3d}, chunks="auto")
-        assert all([is_box(chunk) for chunk in C.fit_transform()])
-
-        C = Chunker({"box": self.BOX3d}, chunks={"lon": 12, "lat": 1, "dpt": 1})
-        assert all([is_box(chunk) for chunk in C.fit_transform()])
-        assert len(C.fit_transform()) == 12
-
-        C = Chunker(
-            {"box": self.BOX3d}, chunks={"lat": 1, "dpt": 1}, chunksize={"lon": 10}
-        )
-        chunks = C.fit_transform()
-        assert all([is_box(chunk) for chunk in chunks])
-        assert chunks[0][1] - chunks[0][0] == 10
-
-        C = Chunker({"box": self.BOX3d}, chunks={"lon": 1, "lat": 12, "dpt": 1})
-        assert all([is_box(chunk) for chunk in C.fit_transform()])
-        assert len(C.fit_transform()) == 12
-
-        C = Chunker(
-            {"box": self.BOX3d}, chunks={"lon": 1, "dpt": 1}, chunksize={"lat": 10}
-        )
-        chunks = C.fit_transform()
-        assert all([is_box(chunk) for chunk in chunks])
-        assert chunks[0][3] - chunks[0][2] == 10
-
-        C = Chunker({"box": self.BOX3d}, chunks={"lon": 1, "lat": 1, "dpt": 12})
-        assert all([is_box(chunk) for chunk in C.fit_transform()])
-        assert len(C.fit_transform()) == 12
-
-        C = Chunker(
-            {"box": self.BOX3d}, chunks={"lon": 1, "lat": 1}, chunksize={"dpt": 10}
-        )
-        chunks = C.fit_transform()
-        assert all([is_box(chunk) for chunk in chunks])
-        assert chunks[0][5] - chunks[0][4] == 10
-
-        C = Chunker({"box": self.BOX3d}, chunks={"lon": 4, "lat": 2, "dpt": 1})
-        assert all([is_box(chunk) for chunk in C.fit_transform()])
-        assert len(C.fit_transform()) == 2 * 4
-
-        C = Chunker({"box": self.BOX3d}, chunks={"lon": 2, "lat": 3, "dpt": 4})
-        assert all([is_box(chunk) for chunk in C.fit_transform()])
-        assert len(C.fit_transform()) == 2 * 3 * 4
-
-        with pytest.raises(ValueError):
-            Chunker({"box": self.BOX3d}, chunks=["lon", 1])
-
-        C = Chunker({"box": self.BOX3d})
-        assert isinstance(C.this_chunker, types.FunctionType) or isinstance(
-            C.this_chunker, types.MethodType
-        )
-
-    def test_chunk_box4d(self):
-        C = Chunker({"box": self.BOX4d})
-        assert all([is_box(chunk) for chunk in C.fit_transform()])
-
-        C = Chunker({"box": self.BOX4d}, chunks="auto")
-        assert all([is_box(chunk) for chunk in C.fit_transform()])
-
-        C = Chunker(
-            {"box": self.BOX4d}, chunks={"lon": 2, "lat": 1, "dpt": 1, "time": 1}
-        )
-        assert all([is_box(chunk) for chunk in C.fit_transform()])
-        assert len(C.fit_transform()) == 2
-
-        C = Chunker(
-            {"box": self.BOX4d},
-            chunks={"lat": 1, "dpt": 1, "time": 1},
-            chunksize={"lon": 10},
-        )
-        chunks = C.fit_transform()
-        assert all([is_box(chunk) for chunk in chunks])
-        assert chunks[0][1] - chunks[0][0] == 10
-
-        C = Chunker(
-            {"box": self.BOX4d}, chunks={"lon": 1, "lat": 2, "dpt": 1, "time": 1}
-        )
-        assert all([is_box(chunk) for chunk in C.fit_transform()])
-        assert len(C.fit_transform()) == 2
-
-        C = Chunker(
-            {"box": self.BOX4d},
-            chunks={"lon": 1, "dpt": 1, "time": 1},
-            chunksize={"lat": 10},
-        )
-        chunks = C.fit_transform()
-        assert all([is_box(chunk) for chunk in chunks])
-        assert chunks[0][3] - chunks[0][2] == 10
-
-        C = Chunker(
-            {"box": self.BOX4d}, chunks={"lon": 1, "lat": 1, "dpt": 2, "time": 1}
-        )
-        assert all([is_box(chunk) for chunk in C.fit_transform()])
-        assert len(C.fit_transform()) == 2
-
-        C = Chunker(
-            {"box": self.BOX4d},
-            chunks={"lon": 1, "lat": 1, "time": 1},
-            chunksize={"dpt": 10},
-        )
-        chunks = C.fit_transform()
-        assert all([is_box(chunk) for chunk in chunks])
-        assert chunks[0][5] - chunks[0][4] == 10
-
-        C = Chunker(
-            {"box": self.BOX4d}, chunks={"lon": 1, "lat": 1, "dpt": 1, "time": 2}
-        )
-        assert all([is_box(chunk) for chunk in C.fit_transform()])
-        assert len(C.fit_transform()) == 2
-
-        C = Chunker(
-            {"box": self.BOX4d},
-            chunks={"lon": 1, "lat": 1, "dpt": 1},
-            chunksize={"time": 5},
-        )
-        chunks = C.fit_transform()
-        assert all([is_box(chunk) for chunk in chunks])
-        assert np.timedelta64(
-            pd.to_datetime(chunks[0][7]) - pd.to_datetime(chunks[0][6]), "D"
-        ) <= np.timedelta64(5, "D")
-
-        with pytest.raises(ValueError):
-            Chunker({"box": self.BOX4d}, chunks=["lon", 1])
-
-        C = Chunker({"box": self.BOX4d})
-        assert isinstance(C.this_chunker, types.FunctionType) or isinstance(
-            C.this_chunker, types.MethodType
-        )
 
 
 def test_format_oneline():
@@ -438,95 +214,3 @@ class Test_argo_split_path:
             assert key in desc
 
 
-class Test_float_wmo():
-
-    def test_init(self):
-        assert isinstance(float_wmo(2901746), float_wmo)
-        assert isinstance(float_wmo(float_wmo(2901746)), float_wmo)
-
-    def test_isvalid(self):
-        assert float_wmo(2901746).isvalid
-        assert not float_wmo(12, errors='ignore').isvalid
-
-    def test_ppt(self):
-        assert isinstance(str(float_wmo(2901746)), str)
-        assert isinstance(repr(float_wmo(2901746)), str)
-
-    def test_comparisons(self):
-        assert float_wmo(2901746) == float_wmo(2901746)
-        assert float_wmo(2901746) != float_wmo(2901745)
-        assert float_wmo(2901746) >= float_wmo(2901746)
-        assert float_wmo(2901746) > float_wmo(2901745)
-        assert float_wmo(2901746) <= float_wmo(2901746)
-        assert float_wmo(2901746) < float_wmo(2901747)
-
-    def test_hashable(self):
-        assert isinstance(hash(float_wmo(2901746)), int)
-
-
-class Test_Registry():
-
-    opts = [(None, 'str'), (['hello', 'world'], str), (None, float_wmo), ([2901746, 4902252], float_wmo)]
-    opts_ids = ["%s, %s" % ((lambda x: 'iterlist' if x is not None else x)(opt[0]), repr(opt[1])) for opt in opts]
-
-    @pytest.mark.parametrize("opts", opts, indirect=False, ids=opts_ids)
-    def test_init(self, opts):
-        assert isinstance(Registry(opts[0], dtype=opts[1]), Registry)
-
-    opts = [(['hello', 'world'], str), ([2901746, 4902252], float_wmo)]
-    opts_ids = ["%s, %s" % ((lambda x: 'iterlist' if x is not None else x)(opt[0]), repr(opt[1])) for opt in opts]
-
-    @pytest.mark.parametrize("opts", opts, indirect=False, ids=opts_ids)
-    def test_commit(self, opts):
-        R = Registry(dtype=opts[1])
-        R.commit(opts[0])
-
-    @pytest.mark.parametrize("opts", opts, indirect=False, ids=opts_ids)
-    def test_append(self, opts):
-        R = Registry(dtype=opts[1])
-        R.append(opts[0][0])
-
-    @pytest.mark.parametrize("opts", opts, indirect=False, ids=opts_ids)
-    def test_extend(self, opts):
-        R = Registry(dtype=opts[1])
-        R.append(opts[0])
-
-    @pytest.mark.parametrize("opts", opts, indirect=False, ids=opts_ids)
-    def test_insert(self, opts):
-        R = Registry(opts[0][0], dtype=opts[1])
-        R.insert(0, opts[0][-1])
-        assert R[0] == opts[0][-1]
-
-    @pytest.mark.parametrize("opts", opts, indirect=False, ids=opts_ids)
-    def test_remove(self, opts):
-        R = Registry(opts[0], dtype=opts[1])
-        R.remove(opts[0][0])
-        assert opts[0][0] not in R
-
-    @pytest.mark.parametrize("opts", opts, indirect=False, ids=opts_ids)
-    def test_copy(self, opts):
-        R = Registry(opts[0], dtype=opts[1])
-        assert R == R.copy()
-
-    bad_opts = [(['hello', 12], str), ([2901746, 1], float_wmo)]
-    bad_opts_ids = ["%s, %s" % ((lambda x: 'iterlist' if x is not None else x)(opt[0]), repr(opt[1])) for opt in opts]
-
-    @pytest.mark.parametrize("opts", bad_opts, indirect=False, ids=bad_opts_ids)
-    def test_invalid_dtype(self, opts):
-        with pytest.raises(ValueError):
-            Registry(opts[0][0], dtype=opts[1], invalid='raise').commit(opts[0][-1])
-        with pytest.warns(UserWarning):
-            Registry(opts[0][0], dtype=opts[1], invalid='warn').commit(opts[0][-1])
-        # Raise nothing:
-        Registry(opts[0][0], dtype=opts[1], invalid='ignore').commit(opts[0][-1])
-
-
-@pytest.mark.parametrize("params", [[6901929, None], [6901929, 12]], indirect=False, ids=['float', 'profile'])
-def test_get_coriolis_profile_id(params, mocked_httpserver):
-    with argopy.set_options(cachedir=tempfile.mkdtemp()):
-        assert isinstance(get_coriolis_profile_id(params[0], params[1], api_server=mocked_server_address), pd.core.frame.DataFrame)
-
-@pytest.mark.parametrize("params", [[6901929, None], [6901929, 12]], indirect=False, ids=['float', 'profile'])
-def test_get_ea_profile_page(params, mocked_httpserver):
-    with argopy.set_options(cachedir=tempfile.mkdtemp()):
-        assert is_list_of_strings(get_ea_profile_page(params[0], params[1], api_server=mocked_server_address))
