+++ argopy/tests/test_utils_caching.py
@@ -1,0 +1,35 @@
+import os
+import pandas as pd
+import argopy
+from argopy import DataFetcher as ArgoDataFetcher
+from utils import (
+    requires_gdac,
+)
+import tempfile
+
+
+@requires_gdac
+def test_clear_cache():
+    ftproot, flist = argopy.tutorial.open_dataset("gdac")
+    with tempfile.TemporaryDirectory() as cachedir:
+        with argopy.set_options(cachedir=cachedir):
+            loader = ArgoDataFetcher(src="gdac", ftp=ftproot, cache=True).profile(2902696, 12)
+            loader.to_xarray()
+            argopy.clear_cache()
+            assert os.path.exists(cachedir) is True
+            assert len(os.listdir(cachedir)) == 0
+
+
+@requires_gdac
+def test_lscache():
+    ftproot, flist = argopy.tutorial.open_dataset("gdac")
+    with tempfile.TemporaryDirectory() as cachedir:
+        with argopy.set_options(cachedir=cachedir):
+            loader = ArgoDataFetcher(src="gdac", ftp=ftproot, cache=True).profile(2902696, 12)
+            loader.to_xarray()
+            result = argopy.utilities.lscache(cache_path=cachedir, prt=True)
+            assert isinstance(result, str)
+
+            result = argopy.utilities.lscache(cache_path=cachedir, prt=False)
+            assert isinstance(result, pd.DataFrame)
+
