+++ argopy/utilities.py
@@ -12,12 +12,10 @@ import warnings
 import urllib
 import json
 import collections
-from collections import UserList
 import copy
 from functools import reduce, wraps
 from packaging import version
 import logging
-from abc import ABC, abstractmethod
 from urllib.parse import urlparse
 from typing import Union
 import inspect
@@ -80,117 +78,6 @@ except ImportError:
 path2assets = importlib.util.find_spec('argopy.static.assets').submodule_search_locations[0]
 
 log = logging.getLogger("argopy.utilities")
-
-
-def clear_cache(fs=None):
-    """ Delete argopy cache folder content """
-    if os.path.exists(OPTIONS["cachedir"]):
-        # shutil.rmtree(OPTIONS["cachedir"])
-        for filename in os.listdir(OPTIONS["cachedir"]):
-            file_path = os.path.join(OPTIONS["cachedir"], filename)
-            try:
-                if os.path.isfile(file_path) or os.path.islink(file_path):
-                    os.unlink(file_path)
-                elif os.path.isdir(file_path):
-                    shutil.rmtree(file_path)
-            except Exception as e:
-                print("Failed to delete %s. Reason: %s" % (file_path, e))
-        if fs:
-            fs.clear_cache()
-
-
-def lscache(cache_path: str = "", prt=True):
-    """ Decode and list cache folder content
-
-        Parameters
-        ----------
-        cache_path: str
-        prt: bool, default=True
-            Return a printable string or a :class:`pandas.DataFrame`
-
-        Returns
-        -------
-        str or :class:`pandas.DataFrame`
-    """
-    from datetime import datetime
-    import math
-    summary = []
-
-    cache_path = OPTIONS['cachedir'] if cache_path == '' else cache_path
-    apath = os.path.abspath(cache_path)
-    log.debug("Listing cache content at: %s" % cache_path)
-
-    def convert_size(size_bytes):
-        if size_bytes == 0:
-            return "0B"
-        size_name = ("B", "KB", "MB", "GB", "TB", "PB", "EB", "ZB", "YB")
-        i = int(math.floor(math.log(size_bytes, 1024)))
-        p = math.pow(1024, i)
-        s = round(size_bytes / p, 2)
-        return "%s %s" % (s, size_name[i])
-
-    cached_files = []
-    fn = os.path.join(apath, "cache")
-    if os.path.exists(fn):
-        with open(fn, "rb") as f:
-            loaded_cached_files = pickle.load(f)  # nosec B301 because files controlled internally
-            for c in loaded_cached_files.values():
-                if isinstance(c["blocks"], list):
-                    c["blocks"] = set(c["blocks"])
-            cached_files.append(loaded_cached_files)
-    else:
-        raise FileSystemHasNoCache("No fsspec cache system at: %s" % apath)
-
-    cached_files = cached_files or [{}]
-    cached_files = cached_files[-1]
-
-    N_FILES = len(cached_files)
-    TOTAL_SIZE = 0
-    for cfile in cached_files:
-        path = os.path.join(apath, cached_files[cfile]['fn'])
-        TOTAL_SIZE += os.path.getsize(path)
-
-    summary.append("%s %s" % ("=" * 20, "%i files in fsspec cache folder (%s)" % (N_FILES, convert_size(TOTAL_SIZE))))
-    summary.append("lscache %s" % os.path.sep.join([apath, ""]))
-    summary.append("=" * 20)
-
-    listing = {'fn': [], 'size': [], 'time': [], 'original': [], 'uid': [], 'blocks': []}
-    for cfile in cached_files:
-        summary.append("- %s" % cached_files[cfile]['fn'])
-        listing['fn'].append(cached_files[cfile]['fn'])
-
-        path = os.path.join(cache_path, cached_files[cfile]['fn'])
-        summary.append("\t%8s: %s" % ('SIZE', convert_size(os.path.getsize(path))))
-        listing['size'].append(os.path.getsize(path))
-
-        key = 'time'
-        ts = cached_files[cfile][key]
-        tsf = pd.to_datetime(datetime.fromtimestamp(ts)).strftime("%c")
-        summary.append("\t%8s: %s (%s)" % (key, tsf, ts))
-        listing['time'].append(pd.to_datetime(datetime.fromtimestamp(ts)))
-
-        if version.parse(fsspec.__version__) > version.parse("0.8.7"):
-            key = 'original'
-            summary.append("\t%8s: %s" % (key, cached_files[cfile][key]))
-            listing[key].append(cached_files[cfile][key])
-
-        key = 'uid'
-        summary.append("\t%8s: %s" % (key, cached_files[cfile][key]))
-        listing[key].append(cached_files[cfile][key])
-
-        key = 'blocks'
-        summary.append("\t%8s: %s" % (key, cached_files[cfile][key]))
-        listing[key].append(cached_files[cfile][key])
-
-    summary.append("=" * 20)
-    summary = "\n".join(summary)
-    if prt:
-        # Return string to be printed:
-        return summary
-    else:
-        # Return dataframe listing:
-        # log.debug(summary)
-        return pd.DataFrame(listing)
 
 
 def get_sys_info():
@@ -620,276 +507,6 @@ def linear_interpolation_remap(
     ).coords[output_dim]
     return remapped
 
-
-class Chunker:
-    """ To chunk fetcher requests """
-
-    # Default maximum chunks size for all possible request parameters
-    default_chunksize = {
-        "box": {
-            "lon": 20,  # degree
-            "lat": 20,  # degree
-            "dpt": 500,  # meters/db
-            "time": 3 * 30,
-        },  # Days
-        "wmo": {"wmo": 5, "cyc": 100},  # Nb of floats
-    }  # Nb of cycles
-
-    def __init__(self, request: dict, chunks: str = "auto", chunksize: dict = {}):
-        """ Create a request Chunker
-
-        Allow to easily split an access point request into chunks
-
-        Parameters
-        ----------
-        request: dict
-            Access point request to be chunked. One of the following:
-
-            - {'box': [lon_min, lon_max, lat_min, lat_max, dpt_min, dpt_max, time_min, time_max]}
-            - {'box': [lon_min, lon_max, lat_min, lat_max, dpt_min, dpt_max]}
-            - {'wmo': [wmo1, wmo2, ...], 'cyc': [0,1, ...]}
-        chunks: 'auto' or dict
-            Dictionary with request access point as keys and number of chunks to create as values.
-
-            Eg: {'wmo':10} will create a maximum of 10 chunks along WMOs.
-        chunksize: dict, optional
-            Dictionary with request access point as keys and chunk size as values (used as maximum values in
-            'auto' chunking).
-
-            Eg: {'wmo': 5} will create chunks with as many as 5 WMOs each.
-
-        """
-        self.request = request
-
-        if "box" in self.request:
-            is_box(self.request["box"])
-            if len(self.request["box"]) == 8:
-                self.this_chunker = self._chunker_box4d
-            elif len(self.request["box"]) == 6:
-                self.this_chunker = self._chunker_box3d
-        elif "wmo" in self.request:
-            self.this_chunker = self._chunker_wmo
-        else:
-            raise InvalidFetcherAccessPoint(
-                "'%s' not valid access point" % ",".join(self.request.keys())
-            )
-
-        default = self.default_chunksize[[k for k in self.request.keys()][0]]
-        if len(chunksize) == 0:  # chunksize = {}
-            chunksize = default
-        if not isinstance(chunksize, collectionsAbc.Mapping):
-            raise ValueError("chunksize must be mappable")
-        else:  # merge with default:
-            chunksize = {**default, **chunksize}
-        self.chunksize = collections.OrderedDict(sorted(chunksize.items()))
-
-        default = {k: "auto" for k in self.chunksize.keys()}
-        if chunks == "auto":  # auto for all
-            chunks = default
-        elif len(chunks) == 0:  # chunks = {}, i.e. chunk=1 for all
-            chunks = {k: 1 for k in self.request}
-        if not isinstance(chunks, collectionsAbc.Mapping):
-            raise ValueError("chunks must be 'auto' or mappable")
-        chunks = {**default, **chunks}
-        self.chunks = collections.OrderedDict(sorted(chunks.items()))
-
-    def _split(self, lst, n=1):
-        """Yield successive n-sized chunks from lst"""
-        for i in range(0, len(lst), n):
-            yield lst[i: i + n]
-
-    def _split_list_bychunknb(self, lst, n=1):
-        """Split list in n-imposed chunks of similar size
-            The last chunk may contain less element than the others, depending on the size of the list.
-        """
-        res = []
-        s = int(np.floor_divide(len(lst), n))
-        for i in self._split(lst, s):
-            res.append(i)
-        if len(res) > n:
-            res[n - 1::] = [reduce(lambda i, j: i + j, res[n - 1::])]
-        return res
-
-    def _split_list_bychunksize(self, lst, max_size=1):
-        """Split list in chunks of imposed size
-            The last chunk may contain less element than the others, depending on the size of the list.
-        """
-        res = []
-        for i in self._split(lst, max_size):
-            res.append(i)
-        return res
-
-    def _split_box(self, large_box, n=1, d="x"):  # noqa: C901
-        """Split a box domain in one direction in n-imposed equal chunks """
-        if d == "x":
-            i_left, i_right = 0, 1
-        if d == "y":
-            i_left, i_right = 2, 3
-        if d == "z":
-            i_left, i_right = 4, 5
-        if d == "t":
-            i_left, i_right = 6, 7
-        if n == 1:
-            return [large_box]
-        boxes = []
-        if d in ["x", "y", "z"]:
-            n += 1  # Required because we split in linspace
-            bins = np.linspace(large_box[i_left], large_box[i_right], n)
-            for ii, left in enumerate(bins):
-                if ii < len(bins) - 1:
-                    right = bins[ii + 1]
-                    this_box = large_box.copy()
-                    this_box[i_left] = left
-                    this_box[i_right] = right
-                    boxes.append(this_box)
-        elif "t" in d:
-            dates = pd.to_datetime(large_box[i_left: i_right + 1])
-            date_bounds = [
-                d.strftime("%Y%m%d%H%M%S")
-                for d in pd.date_range(dates[0], dates[1], periods=n + 1)
-            ]
-            for i1, i2 in zip(np.arange(0, n), np.arange(1, n + 1)):
-                left, right = date_bounds[i1], date_bounds[i2]
-                this_box = large_box.copy()
-                this_box[i_left] = left
-                this_box[i_right] = right
-                boxes.append(this_box)
-        return boxes
-
-    def _split_this_4Dbox(self, box, nx=1, ny=1, nz=1, nt=1):
-        box_list = []
-        split_x = self._split_box(box, n=nx, d="x")
-        for bx in split_x:
-            split_y = self._split_box(bx, n=ny, d="y")
-            for bxy in split_y:
-                split_z = self._split_box(bxy, n=nz, d="z")
-                for bxyz in split_z:
-                    split_t = self._split_box(bxyz, n=nt, d="t")
-                    for bxyzt in split_t:
-                        box_list.append(bxyzt)
-        return box_list
-
-    def _split_this_3Dbox(self, box, nx=1, ny=1, nz=1):
-        box_list = []
-        split_x = self._split_box(box, n=nx, d="x")
-        for bx in split_x:
-            split_y = self._split_box(bx, n=ny, d="y")
-            for bxy in split_y:
-                split_z = self._split_box(bxy, n=nz, d="z")
-                for bxyz in split_z:
-                    box_list.append(bxyz)
-        return box_list
-
-    def _chunker_box4d(self, request, chunks, chunks_maxsize):  # noqa: C901
-        BOX = request["box"]
-        n_chunks = chunks
-        for axis, n in n_chunks.items():
-            if n == "auto":
-                if axis == "lon":
-                    Lx = BOX[1] - BOX[0]
-                    if Lx > chunks_maxsize["lon"]:  # Max box size in longitude
-                        n_chunks["lon"] = int(
-                            np.ceil(np.divide(Lx, chunks_maxsize["lon"]))
-                        )
-                    else:
-                        n_chunks["lon"] = 1
-                if axis == "lat":
-                    Ly = BOX[3] - BOX[2]
-                    if Ly > chunks_maxsize["lat"]:  # Max box size in latitude
-                        n_chunks["lat"] = int(
-                            np.ceil(np.divide(Ly, chunks_maxsize["lat"]))
-                        )
-                    else:
-                        n_chunks["lat"] = 1
-                if axis == "dpt":
-                    Lz = BOX[5] - BOX[4]
-                    if Lz > chunks_maxsize["dpt"]:  # Max box size in depth
-                        n_chunks["dpt"] = int(
-                            np.ceil(np.divide(Lz, chunks_maxsize["dpt"]))
-                        )
-                    else:
-                        n_chunks["dpt"] = 1
-                if axis == "time":
-                    Lt = np.timedelta64(
-                        pd.to_datetime(BOX[7]) - pd.to_datetime(BOX[6]), "D"
-                    )
-                    MaxLen = np.timedelta64(chunks_maxsize["time"], "D")
-                    if Lt > MaxLen:  # Max box size in time
-                        n_chunks["time"] = int(np.ceil(np.divide(Lt, MaxLen)))
-                    else:
-                        n_chunks["time"] = 1
-
-        boxes = self._split_this_4Dbox(
-            BOX,
-            nx=n_chunks["lon"],
-            ny=n_chunks["lat"],
-            nz=n_chunks["dpt"],
-            nt=n_chunks["time"],
-        )
-        return {"chunks": sorted(n_chunks), "values": boxes}
-
-    def _chunker_box3d(self, request, chunks, chunks_maxsize):
-        BOX = request["box"]
-        n_chunks = chunks
-        for axis, n in n_chunks.items():
-            if n == "auto":
-                if axis == "lon":
-                    Lx = BOX[1] - BOX[0]
-                    if Lx > chunks_maxsize["lon"]:  # Max box size in longitude
-                        n_chunks["lon"] = int(
-                            np.floor_divide(Lx, chunks_maxsize["lon"])
-                        )
-                    else:
-                        n_chunks["lon"] = 1
-                if axis == "lat":
-                    Ly = BOX[3] - BOX[2]
-                    if Ly > chunks_maxsize["lat"]:  # Max box size in latitude
-                        n_chunks["lat"] = int(
-                            np.floor_divide(Ly, chunks_maxsize["lat"])
-                        )
-                    else:
-                        n_chunks["lat"] = 1
-                if axis == "dpt":
-                    Lz = BOX[5] - BOX[4]
-                    if Lz > chunks_maxsize["dpt"]:  # Max box size in depth
-                        n_chunks["dpt"] = int(
-                            np.floor_divide(Lz, chunks_maxsize["dpt"])
-                        )
-                    else:
-                        n_chunks["dpt"] = 1
-                # if axis == 'time':
-                #     Lt = np.timedelta64(pd.to_datetime(BOX[5]) - pd.to_datetime(BOX[4]), 'D')
-                #     MaxLen = np.timedelta64(chunks_maxsize['time'], 'D')
-                #     if Lt > MaxLen:  # Max box size in time
-                #         n_chunks['time'] = int(np.floor_divide(Lt, MaxLen))
-                #     else:
-                #         n_chunks['time'] = 1
-        boxes = self._split_this_3Dbox(
-            BOX, nx=n_chunks["lon"], ny=n_chunks["lat"], nz=n_chunks["dpt"]
-        )
-        return {"chunks": sorted(n_chunks), "values": boxes}
-
-    def _chunker_wmo(self, request, chunks, chunks_maxsize):
-        WMO = request["wmo"]
-        n_chunks = chunks
-        if n_chunks["wmo"] == "auto":
-            wmo_grps = self._split_list_bychunksize(WMO, max_size=chunks_maxsize["wmo"])
-        else:
-            n = np.min([n_chunks["wmo"], len(WMO)])
-            wmo_grps = self._split_list_bychunknb(WMO, n=n)
-        n_chunks["wmo"] = len(wmo_grps)
-        return {"chunks": sorted(n_chunks), "values": wmo_grps}
-
-    def fit_transform(self):
-        """ Chunk a fetcher request
-
-        Returns
-        -------
-        list
-        """
-        self._results = self.this_chunker(self.request, self.chunks, self.chunksize)
-        # self.chunks = self._results['chunks']
-        return self._results["values"]
 
 
 def format_oneline(s, max_width=65):
@@ -1365,258 +982,6 @@ def argo_split_path(this_path):  # noqa C901
     output['path'] = output['path'].replace(output['origin'], '')
 
     return dict(sorted(output.items()))
-
-
-class RegistryItem(ABC):
-    """Prototype for possible custom items in a Registry"""
-    @property
-    @abstractmethod
-    def value(self):
-        raise NotImplementedError("Not implemented")
-
-    @property
-    @abstractmethod
-    def isvalid(self, item):
-        raise NotImplementedError("Not implemented")
-
-    @abstractmethod
-    def __str__(self):
-        raise NotImplementedError("Not implemented")
-
-    @abstractmethod
-    def __repr__(self):
-        raise NotImplementedError("Not implemented")
-
-
-class float_wmo(RegistryItem):
-    """Argo float WMO number object"""
-
-    def __init__(self, WMO_number, errors='raise'):
-        """Create an Argo float WMO number object
-
-        Parameters
-        ----------
-        WMO_number: object
-            Anything that could be casted as an integer
-        errors: {'raise', 'warn', 'ignore'}
-            Possibly raises a ValueError exception or UserWarning, otherwise fails silently if WMO_number is not valid
-
-        Returns
-        -------
-        :class:`argopy.utilities.float_wmo`
-        """
-        self.errors = errors
-        if isinstance(WMO_number, float_wmo):
-            item = WMO_number.value
-        else:
-            item = check_wmo(WMO_number, errors=self.errors)[0]  # This will automatically validate item
-        self.item = item
-
-    @property
-    def isvalid(self):
-        """Check if WMO number is valid"""
-        return is_wmo(self.item, errors=self.errors)
-        # return True  # Because it was checked at instantiation
-
-    @property
-    def value(self):
-        """Return WMO number as in integer"""
-        return int(self.item)
-
-    def __str__(self):
-        # return "%s" % check_wmo(self.item)[0]
-        return "%s" % self.item
-
-    def __repr__(self):
-        return f"WMO({self.item})"
-
-    def __check_other__(self, other):
-        return check_wmo(other)[0] if type(other) is not float_wmo else other.item
-
-    def __eq__(self, other):
-        return self.item.__eq__(self.__check_other__(other))
-
-    def __ne__(self, other):
-        return self.item.__ne__(self.__check_other__(other))
-
-    def __gt__(self, other):
-        return self.item.__gt__(self.__check_other__(other))
-
-    def __lt__(self, other):
-        return self.item.__lt__(self.__check_other__(other))
-
-    def __ge__(self, other):
-        return self.item.__ge__(self.__check_other__(other))
-
-    def __le__(self, other):
-        return self.item.__le__(self.__check_other__(other))
-
-    def __hash__(self):
-        return hash(self.item)
-
-
-class Registry(UserList):
-    """A list manager can that validate item type
-
-    Examples
-    --------
-    You can commit new entry to the registry, one by one:
-
-        >>> R = Registry(name='file')
-        >>> R.commit('meds/4901105/profiles/D4901105_017.nc')
-        >>> R.commit('aoml/1900046/profiles/D1900046_179.nc')
-
-    Or with a list:
-
-        >>> R = Registry(name='My floats', dtype='wmo')
-        >>> R.commit([2901746, 4902252])
-
-    And also at instantiation time (name and dtype are optional):
-
-        >>> R = Registry([2901746, 4902252], name='My floats', dtype=float_wmo)
-
-    Registry can be used like a list.
-
-    It is iterable:
-
-        >>> for wmo in R:
-        >>>     print(wmo)
-
-    It has a ``len`` property:
-
-        >>> len(R)
-
-    It can be checked for values:
-
-        >>> 4902252 in R
-
-    You can also remove items from the registry, again one by one or with a list:
-
-        >>> R.remove('2901746')
-
-    """
-
-    def _complain(self, msg):
-        if self._invalid == 'raise':
-            raise ValueError(msg)
-        elif self._invalid == 'warn':
-            warnings.warn(msg)
-        else:
-            log.debug(msg)
-
-    def _str(self, item):
-        is_valid = isinstance(item, str)
-        if not is_valid:
-            self._complain("%s is not a valid %s" % (str(item), self.dtype))
-        return is_valid
-
-    def _dict(self, item):
-        is_valid = isinstance(item, dict)
-        if not is_valid:
-            self._complain("%s is not a valid %s" % (str(item), self.dtype))
-        return is_valid
-
-    def _wmo(self, item):
-        return item.isvalid
-
-    def __init__(self, initlist=None, name: str = 'unnamed', dtype='str', invalid='raise'):
-        """Create a registry, i.e. a controlled list
-
-        Parameters
-        ----------
-        initlist: list, optional
-            List of values to register
-        name: str, default: 'unnamed'
-            Name of the Registry
-        dtype: :class:`str` or dtype, default: :class:`str`
-            Data type of registry content. Supported values are: 'str', 'wmo', float_wmo
-        invalid: str, default: 'raise'
-            Define what do to when a new item is not valid. Can be 'raise' or 'ignore'
-        """
-        self.name = name
-        self._invalid = invalid
-        if repr(dtype) == "<class 'str'>" or dtype == 'str':
-            self._validator = self._str
-            self.dtype = str
-        elif dtype == float_wmo or str(dtype).lower() == 'wmo':
-            self._validator = self._wmo
-            self.dtype = float_wmo
-        elif repr(dtype) == "<class 'dict'>" or dtype == 'dict':
-            self._validator = self._dict
-            self.dtype = dict
-        elif hasattr(dtype, 'isvalid'):
-            self._validator = dtype.isvalid
-            self.dtype = dtype
-        else:
-            raise ValueError("Unrecognised Registry data type '%s'" % dtype)
-
-        if initlist is not None:
-            initlist = self._process_items(initlist)
-        super().__init__(initlist)
-
-    def __repr__(self):
-        summary = ["<argopy.registry>%s" % str(self.dtype)]
-        summary.append("Name: %s" % self.name)
-        N = len(self.data)
-        msg = "Nitems: %s" % N if N > 1 else "Nitem: %s" % N
-        summary.append(msg)
-        if N > 0:
-            items = [str(item) for item in self.data]
-            # msg = format_oneline("[%s]" % "; ".join(items), max_width=120)
-            msg = "[%s]" % "; ".join(items)
-            summary.append("Content: %s" % msg)
-        return "\n".join(summary)
-
-    def _process_items(self, items):
-        if not isinstance(items, list):
-            items = [items]
-        if self.dtype == float_wmo:
-            items = [float_wmo(item, errors=self._invalid) for item in items]
-        return items
-
-    def commit(self, values):
-        """R.commit(values) -- append values to the end of the registry if not already in"""
-        items = self._process_items(values)
-        for item in items:
-            if item not in self.data and self._validator(item):
-                super().append(item)
-        return self
-
-    def append(self, value):
-        """R.append(value) -- append value to the end of the registry"""
-        items = self._process_items(value)
-        for item in items:
-            if self._validator(item):
-                super().append(item)
-        return self
-
-    def extend(self, other):
-        """R.extend(iterable) -- extend registry by appending elements from the iterable"""
-        self.append(other)
-        return self
-
-    def remove(self, values):
-        """R.remove(valueS) -- remove first occurrence of values."""
-        items = self._process_items(values)
-        for item in items:
-            if item in self.data:
-                super().remove(item)
-        return self
-
-    def insert(self, index, value):
-        """R.insert(index, value) -- insert value before index."""
-        item = self._process_items(value)[0]
-        if self._validator(item):
-            super().insert(index, item)
-        return self
-
-    def __copy__(self):
-        # Called with copy.copy(R)
-        return Registry(copy.copy(self.data), dtype=self.dtype)
-
-    def copy(self):
-        """Return a shallow copy of the registry"""
-        return self.__copy__()
 
 
 def log_argopy_callerstack(level='debug'):
